+++ src/algorithm/centroid_algorithm.c
@@ -41,7 +41,7 @@
 
 static inline void __attribute__((__always_inline__,__gnu_inline__,__nonnull__,__artificial__))
 centroid_run (const VECTOR* vx, const VECTOR* vy,
-              const VECTOR *restrict r, size_t num_anchors, int width __attribute__((__unused__)), int height __attribute__((__unused__)), VECTOR *restrict resx, VECTOR *restrict resy)
+              const VECTOR *restrict r, size_t num_anchors, size_t width __attribute__((__unused__)), size_t height __attribute__((__unused__)), VECTOR *restrict resx, VECTOR *restrict resy)
 {
     // px and py hold the intersections, sx, sy the sum of the intersection
     // coordinates to accumulate for the final result.
