+++ src/algorithm/clurol_algorithm.c
@@ -211,7 +211,7 @@ center_of_mass_eqwp(const Point2dC *pts,const int count,
 static inline void
 __attribute__((__always_inline__,__gnu_inline__,__artificial__,__nonnull__))
 clurol_run(const VECTOR* vx, const VECTOR* vy, const VECTOR *restrict r,
-              size_t no_anchors, int width, int height,
+              size_t no_anchors, size_t width, size_t height,
               VECTOR *restrict resx, VECTOR *restrict resy)
 {
 if (width==height){};
