+++ src/algorithm/geon_algorithm.c
@@ -76,7 +76,7 @@ circle_minimum_circle_containment(size_t mcount, float *mx, float *my,
 
 static inline void __attribute__((__always_inline__,__gnu_inline__,__nonnull__,__artificial__))
 geon_run (const VECTOR* vx, const VECTOR* vy,
-          const VECTOR *restrict r, size_t num_anchors, int width __attribute__((__unused__)), int height __attribute__((__unused__)), VECTOR *restrict resx, VECTOR *restrict resy)
+          const VECTOR *restrict r, size_t num_anchors, size_t width __attribute__((__unused__)), size_t height __attribute__((__unused__)), VECTOR *restrict resx, VECTOR *restrict resy)
 {
     if (num_anchors < 3) return;
